pyproject.toml
[project]
name = "ai-pipelines"
version = "0.1.0"
description = "YAML-driven AI/LLM pipeline executor"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "jsonschema>=4.20",
    "jsonata-python>=0.6",
    "claude-agent-sdk>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_pipelines"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
